requirements.txt
# Core Audio & Signal Processing
numpy>=1.24.0
sounddevice>=0.4.6

# Speech Processing (Conservative Path - Default)
silero-vad>=5.1.0
faster-whisper>=0.9.0

# LLM Inference (Conservative Path - Default)
# For modern path, replace Ollama with: vllm, bitsandbytes
requests>=2.31.0

# Text-to-Speech (Conservative Path - Default)
# For modern path, replace with: torch, torchaudio, coqui-tts
piper-tts>=1.2.0

# Framework & Integration
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

# Development & Testing (Optional - for contributors)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
